Move all *Options to metav1
