+++ pkg/apis/autoscaling/register.go
@@ -19,7 +19,6 @@ package autoscaling
 import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/kubernetes/pkg/api"
 )
 
 // GroupName is the group name use in this package
@@ -49,7 +48,6 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&Scale{},
 		&HorizontalPodAutoscaler{},
 		&HorizontalPodAutoscalerList{},
-		&api.DeleteOptions{},
 	)
 	return nil
 }
