+++ pkg/apis/extensions/register.go
@@ -17,10 +17,8 @@ limitations under the License.
 package extensions
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/apis/autoscaling"
 )
 
@@ -64,11 +62,8 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&ThirdPartyResourceDataList{},
 		&Ingress{},
 		&IngressList{},
-		&api.DeleteOptions{},
 		&ReplicaSet{},
 		&ReplicaSetList{},
-		&metav1.ExportOptions{},
-		&metav1.GetOptions{},
 		&PodSecurityPolicy{},
 		&PodSecurityPolicyList{},
 		&NetworkPolicy{},
