+++ pkg/apis/imagepolicy/register.go
@@ -17,10 +17,8 @@ limitations under the License.
 package imagepolicy
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/kubernetes/pkg/api"
 )
 
 // GroupName is the group name use in this package
@@ -46,10 +44,6 @@ var (
 
 func addKnownTypes(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(SchemeGroupVersion,
-		&api.DeleteOptions{},
-		&metav1.ExportOptions{},
-		&metav1.GetOptions{},
-
 		&ImageReview{},
 	)
 	//	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
