+++ pkg/apis/policy/types.go
@@ -18,7 +18,6 @@ package policy
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/util/intstr"
 )
 
@@ -110,5 +109,5 @@ type Eviction struct {
 
 	// DeleteOptions may be provided
 	// +optional
-	DeleteOptions *api.DeleteOptions
+	DeleteOptions *metav1.DeleteOptions
 }
